+++ foxes/algorithms/downwind/models/point_wakes_calc.py
@@ -15,12 +15,14 @@ class PointWakesCalculation(PointDataModel):
         The extra evaluation models
     emodels_cpars: list of dict
         The calculation parameters for extra models
+    wake_models: list of foxes.core.WakeModel
+        The wake models, default: from algo
 
     :group: algorithms.downwind.models
 
     """
 
-    def __init__(self, point_vars=None, emodels=None, emodels_cpars=None):
+    def __init__(self, point_vars=None, emodels=None, emodels_cpars=None, wake_models=None):
         """
         Constructor.
 
@@ -32,12 +34,15 @@ class PointWakesCalculation(PointDataModel):
             The extra evaluation models
         emodels_cpars: list of dict, optional
             The calculation parameters for extra models
+        wake_models: list of foxes.core.WakeModel, optional
+            The wake models, default: from algo
 
         """
         super().__init__()
         self._pvars = point_vars
         self.emodels = emodels
         self.emodels_cpars = emodels_cpars
+        self.wake_models = wake_models
 
     def initialize(self, algo, verbosity=0):
         """
@@ -90,8 +95,50 @@ class PointWakesCalculation(PointDataModel):
 
         """
         return self.pvars
+    
+    def contribute_to_wake_deltas(
+            self, 
+            algo,
+            mdata, 
+            fdata, 
+            pdata, 
+            states_source_turbine,
+            wmodels,
+            wdeltas,
+        ):
+        """
+        Contribute to wake deltas from source turbines
+        
+        Parameters
+        ----------
+        algo: foxes.core.Algorithm
+            The calculation algorithm
+        mdata: foxes.core.Data
+            The model data
+        fdata: foxes.core.Data
+            The farm data
+        pdata: foxes.core.Data
+            The point data
+        states_source_turbine: numpy.ndarray
+            For each state, one turbine index for the
+            wake causing turbine. Shape: (n_states,)
+        wmodels: list of foxes.core.WakeModel
+            The wake models
+        wdeltas: dict
+            The wake deltas, are being modified ob the fly.
+            Key: Variable name str, for which the
+            wake delta applies, values: numpy.ndarray with
+            shape (n_states, n_points, ...)
+            
+        """
+        wcoos = algo.wake_frame.get_wake_coos(algo, mdata, fdata, pdata, states_source_turbine)
 
-    def calculate(self, algo, mdata, fdata, pdata):
+        for w in wmodels:
+            w.contribute_to_wake_deltas(
+                algo, mdata, fdata, pdata, states_source_turbine, wcoos, wdeltas
+            )
+    
+    def calculate(self, algo, mdata, fdata, pdata, states_source_turbine=None):
         """ "
         The main model calculation.
 
@@ -108,6 +155,10 @@ class PointWakesCalculation(PointDataModel):
             The farm data
         pdata: foxes.core.Data
             The point data
+        states_source_turbine: numpy.ndarray, optional
+            For each state, one turbine index for the
+            wake causing turbine. Shape: (n_states,).
+            Default: include all turbines
 
         Returns
         -------
@@ -118,10 +169,11 @@ class PointWakesCalculation(PointDataModel):
         """
         torder = fdata[FV.ORDER].astype(FC.ITYPE)
         n_order = torder.shape[1]
+        wake_models = algo.wake_models if self.wake_models is None else self.wake_models
 
         wdeltas = {}
         wmodels = []
-        for w in algo.wake_models:
+        for w in wake_models:
             hdeltas = {}
             w.init_wake_deltas(algo, mdata, fdata, pdata, hdeltas)
             if len(set(self.pvars).intersection(hdeltas.keys())):
@@ -129,14 +181,13 @@ class PointWakesCalculation(PointDataModel):
                 wmodels.append(w)
             del hdeltas
 
-        for oi in range(n_order):
-            o = torder[:, oi]
-            wcoos = algo.wake_frame.get_wake_coos(algo, mdata, fdata, pdata, o)
-
-            for w in wmodels:
-                w.contribute_to_wake_deltas(
-                    algo, mdata, fdata, pdata, o, wcoos, wdeltas
-                )
+        if states_source_turbine is None:
+            for oi in range(n_order):
+                o = torder[:, oi]
+                self.contribute_to_wake_deltas(algo, mdata, fdata, pdata, o, wmodels, wdeltas)
+        else:
+            self.contribute_to_wake_deltas(algo, mdata, fdata, pdata, states_source_turbine, 
+                                           wmodels, wdeltas)
 
         amb_res = {v: pdata[FV.var2amb[v]] for v in wdeltas}
         for w in wmodels: